fix(MessageSender): 方法overload造成的混淆

+++ src/main/java/com/insanexs/message/MessageSenderFactory.java
@@ -7,6 +7,6 @@ package com.insanexs.message;
  */
 public interface MessageSenderFactory {
 
-    MessageSender createMessageSender(String protocol, int timeout);
+    MessageSender createMessageSender(String protocol, long timeout);
 
 }
