+++ src/main/java/com/insanexs/message/protocol/modbusTCP/ModbusTCPMessageSenderFactory.java
@@ -1,11 +1,7 @@
 package com.insanexs.message.protocol.modbusTCP;
 
 import com.insanexs.message.MessageSender;
-import com.insanexs.message.MessageSenderFactory;
 import com.insanexs.message.PooledMessageSenderFactory;
-
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * @Author: xieshang
@@ -14,13 +10,13 @@ import java.util.regex.Pattern;
  */
 public class ModbusTCPMessageSenderFactory implements PooledMessageSenderFactory {
     @Override
-    public MessageSender createMessageSender(String protocol, int timeout) {
+    public MessageSender createMessageSender(String protocol, long timeout) {
         return new ModbusTCPMessageSender(protocol, timeout);
     }
 
 
     @Override
-    public MessageSender createPooledMessageSender(String protocol, int timeout) {
+    public MessageSender createPooledMessageSender(String protocol, long timeout) {
         return null;
     }
 }
